requirements.txt
# General
Django
django-htmx
django-bootstrap5
django-htmx
django-bootstrap-v5
numerize

# Testing
pytest-django
factory_boy
pytest-cov

# Other
pre-commit
pylint
django-extensions
